+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -   Added measure tools
 -   Added layout to map filters
 -   Added state, sensor type, and network name services and filter selections
+-   Added supplementary layers to toggle
 
 ### Changed